fix: 새로고침 방지 훅 적용
