+++ src/app/reservation/[id]/write/payments/page.tsx
@@ -5,12 +5,17 @@ import { postPayment } from '@/services/billing.service';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useRef } from 'react';
 import { BeatLoader } from 'react-spinners';
+import usePreventScroll from '@/hooks/usePreventScroll';
+import usePreventRefresh from '@/hooks/usePreventRefresh';
 
 const Page = () => {
   const router = useRouter();
   const searchParams = useSearchParams();
   const pathname = usePathname();
   const isInitiated = useRef(false);
+
+  usePreventRefresh();
+  usePreventScroll();
 
   const callPaymentConfirmation = async () => {
     try {
@@ -35,16 +40,6 @@ const Page = () => {
     }
     isInitiated.current = true;
     callPaymentConfirmation();
-  }, []);
-
-  useEffect(() => {
-    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
-      e.preventDefault();
-    };
-    window.addEventListener('beforeunload', handleBeforeUnload);
-    return () => {
-      window.removeEventListener('beforeunload', handleBeforeUnload);
-    };
   }, []);
 
   return (
