+++ src/hooks/usePreventRefresh.tsx
@@ -1,0 +1,15 @@
+import { useEffect } from 'react';
+
+const usePreventRefresh = () => {
+  useEffect(() => {
+    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
+      e.preventDefault();
+    };
+    window.addEventListener('beforeunload', handleBeforeUnload);
+    return () => {
+      window.removeEventListener('beforeunload', handleBeforeUnload);
+    };
+  }, []);
+};
+
+export default usePreventRefresh;
